:pencil2::new: Window::enumerate()
